pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fireaudit"
version = "0.1.0"
description = "Firewall configuration auditing tool with multi-vendor support and compliance framework mapping"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "jinja2>=3.1",
    "click>=8.1",
    "rich>=13.0",
    "questionary>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.8",
    "pyinstaller>=6.0",
]

[project.scripts]
fireaudit = "fireaudit.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fireaudit*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
